major updates in processing

+++ commons/src/main/java/be/fooda/backend/commons/model/template/order/request/FoodaOrderProductReq.java
@@ -1,5 +1,3 @@
-// OrderedProduct.java
-
 package be.fooda.backend.commons.model.template.order.request;
 
 import java.math.BigDecimal;
@@ -12,7 +10,7 @@ import lombok.*;
 public class FoodaOrderProductReq {
     private Long productId;
     private String name;
-    private Long quantity;
+    private Integer quantity;
     private BigDecimal price;
     private FoodaOrderVariantReq variant;
     private String type;
